break cycles in /recipe/
