+++ src/runtime/manifest.ts
@@ -26,7 +26,8 @@ import {Handle} from './recipe/handle.js';
 import {Particle} from './recipe/particle.js';
 import {Slot} from './recipe/slot.js';
 import {HandleConnection} from './recipe/handle-connection.js';
-import {RecipeUtil, connectionMatchesHandleDirection} from './recipe/recipe-util.js';
+import {RecipeUtil} from './recipe/recipe-util.js';
+import {connectionMatchesHandleDirection} from './recipe/direction-util.js';
 import {Recipe, RequireSection} from './recipe/recipe.js';
 import {Search} from './recipe/search.js';
 import {TypeChecker} from './recipe/type-checker.js';
