+++ src/runtime/recipe/recipe-util.ts
@@ -12,6 +12,7 @@ import {assert} from '../../platform/assert-web.js';
 import {ParticleSpec, HandleConnectionSpec} from '../particle-spec.js';
 import {InterfaceType} from '../type.js';
 
+import {connectionMatchesHandleDirection} from './direction-util.js';
 import {HandleConnection} from './handle-connection.js';
 import {Direction} from '../manifest-ast-nodes.js';
 import {Handle} from './handle.js';
@@ -39,37 +40,6 @@ export function reverseDirection(direction: Direction): Direction {
       throw new Error(`Bad direction ${direction}`);
   }
 }
-
-export function connectionMatchesHandleDirection(connectionDirection: Direction, handleDirection: Direction): boolean {
-  return acceptedDirections(connectionDirection).includes(handleDirection);
-}
-
-export function acceptedDirections(direction: Direction): Direction[] {
-  // @param direction: the direction of a handleconnection.
-  // @return acceptedDirections: the list of directions a handle can have that
-  // are allowed with this handle connection.
-  //
-  switch (direction) {
-    case 'any':
-      return ['any', 'reads', 'writes', 'reads writes', 'hosts', '`consumes', '`provides'];
-    case 'reads':
-      return ['any', 'reads', 'reads writes', 'hosts', '`consumes'];
-    case 'writes':
-      return ['any', 'writes', 'reads writes', '`provides'];
-    case 'reads writes':
-      return ['any', 'reads writes'];
-    case 'hosts':
-      return ['any', 'hosts'];
-    case '`consumes':
-      return ['any', '`consumes'];
-    case '`provides':
-      return ['any', '`provides'];
-    default:
-      // Catch nulls and unsafe values from javascript.
-      throw new Error(`Bad direction ${direction}`);
-  }
-}
-
 
 class Shape {
   recipe: Recipe;
