+++ src/runtime/recipe/util.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright (c) 2020 Google Inc. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * Code distributed by Google as part of this project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {Recipe} from './recipe.js';
+
+export function isRequireSection(recipe: Recipe) {
+  return 'parent' in recipe;
+}
